+++ CHANGELOG.md
@@ -19,7 +19,7 @@
   addLabelTo (-> addAnnotationTo).
 - Removed deprecated WmsControls.setPlayButton, setStopButton, setPlayStopButton, setPauseButton
   (-> setSinglePlay, setSinglePlayStop, setPlayStop, setPlayPauseStop).
-
+- Removed special case target for grand staff.
 
 ## [6.4.1] - 2026-03-22
 ### Fixed